+++ src/test/java/com/vonage/client/AbstractMethodTest.java
@@ -20,7 +20,10 @@ import static com.vonage.client.TestUtils.*;
 import com.vonage.client.auth.*;
 import com.vonage.client.auth.hashutils.HashType;
 import org.apache.commons.codec.binary.Base64;
-import org.apache.http.*;
+import org.apache.http.HttpEntity;
+import org.apache.http.HttpEntityEnclosingRequest;
+import org.apache.http.HttpResponse;
+import org.apache.http.ProtocolVersion;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.methods.RequestBuilder;
@@ -34,7 +37,10 @@ import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.*;
 import org.mockito.ArgumentCaptor;
 import static org.mockito.Mockito.*;
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.net.InetSocketAddress;
 import java.net.SocketTimeoutException;
 import java.nio.charset.StandardCharsets;
@@ -119,6 +125,9 @@ public class AbstractMethodTest {
         AuthCollection mockAuthMethods = mock(AuthCollection.class);
         mockAuthMethod = mock(AuthMethod.class);
         mockHttpClient = mock(CloseableHttpClient.class);
+        var httpConfig = HttpConfig.builder().build();
+        when(mockWrapper.getHttpConfig()).thenReturn(httpConfig);
+        when(mockAuthMethod.getSortKey()).thenReturn(0);
         when(mockAuthMethods.getAcceptableAuthMethod(any())).thenReturn(mockAuthMethod);
         when(mockWrapper.getHttpClient()).thenReturn(mockHttpClient);
         when(mockHttpClient.execute(any(HttpUriRequest.class))).thenReturn(basicResponse);
@@ -157,6 +166,22 @@ public class AbstractMethodTest {
         String result = method.execute("url");
         String expected = "User-Agent: "+userAgent;
         assertEquals(expected, result);
+    }
+
+    @Test
+    public void testCustomHeaderDoesNotOverrideUserAgent() {
+        var overridenUa = "my-custom-agent Should not be possible";
+        var cm = new ConcreteMethod(new HttpWrapper(HttpConfig.builder()
+                    .addRequestHeader("X-Correlation-Id", "aaaa-bbbb-cccc-dddd")
+                    .addRequestHeader("User-Agent", overridenUa)
+                .build(), new NoAuthMethod()
+        ));
+        var request = cm.createFullHttpRequest("https://example.com");
+        var userAgentHeaders = request.getHeaders("User-Agent");
+        assertNotNull(userAgentHeaders);
+        assertEquals(1, userAgentHeaders.length);
+        assertNotEquals(overridenUa, userAgentHeaders[0].getValue());
+        assertEquals("aaaa-bbbb-cccc-dddd", request.getFirstHeader("X-Correlation-Id").getValue());
     }
 
     @Test