+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# [9.2.0] - 2025-04-30
+- Added support for setting additional request headers in `HttpConfig`
+- Allow setting `HttpClient` and `HttpConfig` on `HttpWrapper`
+- Bumped Jackson version to 2.19.0
+
 # [9.1.0] - 2025-04-22
 - Added custom HTTP requests support via `CustomClient` (see [README](README.md#custom-requests) for details)